Rename some files to make the peoples happey
